chore/move articlePreview to components
