+++ web/src/components/ArticlesCell/ArticlesCell.tsx
@@ -2,7 +2,7 @@ import type { ArticlesQuery } from 'types/graphql'
 
 import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 
-import ArticlePreview from '../Article/ArticlePreview'
+import ArticlePreview from '../Article/components/ArticlePreview'
 import Skeleton from '../Skeleton/Skeleton'
 
 export const QUERY = gql`
